Replace the VideoEncoderStream with a new thread-safe class.
